# api/docs/swagger.yaml
definitions:
  booking.AccessCoach:
    properties:
      booking_coach_id:
        type: string
      date:
        type: string
    type: object
  booking.AccessGroup:
    properties:
      booking_group_id:
        type: string
      date:
        type: string
    type: object
  booking.AccessPersonal:
    properties:
      booking_personal_id:
        type: string
      date:
        type: string
    type: object
...
  booking.CreateAccessCoachRequest:
    properties:
      access_coach:
        $ref: '#/definitions/booking.AccessCoach'
    type: object
  booking.CreateAccessGroupRequest:
    properties:
      access_group:
        $ref: '#/definitions/booking.AccessGroup'
    type: object
  booking.CreateAccessPersonalRequest:
    properties:
      access_personal:
        $ref: '#/definitions/booking.AccessPersonal'
    type: object
  booking.ListAccessCoachResponse:
    properties:
      access_coach:
        items:
          $ref: '#/definitions/booking.AccessCoach'
        type: array
    type: object
  booking.ListAccessGroupResponse:
    properties:
      access_group:
        items:
          $ref: '#/definitions/booking.AccessGroup'
        type: array
    type: object
  booking.ListAccessPersonalResponse:
    properties:
      access_personal:
        items:
          $ref: '#/definitions/booking.AccessPersonal'
        type: array
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /v1/access/coach:
    post:
      consumes:
      - application/json
      description: Create a new access coach record with the provided details.
      parameters:
      - description: Access Coach details
        in: body
        name: access_coach
        required: true
        schema:
          $ref: '#/definitions/booking.CreateAccessCoachRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/booking.AccessCoach'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Access Coach record
      tags:
      - Access
  /v1/access/coach/{booking_coach_id}:
    get:
      consumes:
      - application/json
      description: Get a list of access coach records for a specific booking.
      parameters:
      - description: Booking Coach ID
        in: path
        name: booking_coach_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListAccessCoachResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Access Coach records
      tags:
      - Access
  /v1/access/group:
    post:
      consumes:
      - application/json
      description: Create a new access group record with the provided details.
      parameters:
      - description: Access Group details
        in: body
        name: access_group
        required: true
        schema:
          $ref: '#/definitions/booking.CreateAccessGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/booking.AccessGroup'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Access Group record
      tags:
      - Access
  /v1/access/group/{booking_group_id}:
    get:
      consumes:
      - application/json
      description: Get a list of access group records for a specific booking.
      parameters:
      - description: Booking Group ID
        in: path
        name: booking_group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListAccessGroupResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Access Group records
      tags:
      - Access
  /v1/access/personal:
    post:
      consumes:
      - application/json
      description: Create a new access personal record with the provided details.
      parameters:
      - description: Access Personal details
        in: body
        name: access_personal
        required: true
        schema:
          $ref: '#/definitions/booking.CreateAccessPersonalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/booking.AccessPersonal'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Access Personal record
      tags:
      - Access
  /v1/access/personal/{booking_personal_id}:
    get:
      consumes:
      - application/json
      description: Get a list of access personal records for a specific booking.
      parameters:
      - description: Booking Personal ID
        in: path
        name: booking_personal_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListAccessPersonalResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Access Personal records
      tags:
      - Access
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"